* 5l/: update to latest 5l from https://github.com/0intro/plan9/ backport all the modifications (remove thumb stuff, add -f, add -L and libpath, add -H0 and -H6, ...).
Goal is to be able to compiler/link the bcm/ kernel!
with the not-updated kencc version I was getting just
a blue screen. Calls to malloc were failing for unknown reasons.

The modif to 5l are not enough I think, I also need to update
cc/ and 5c/ but it's an important part I think.